Some updates
website: add/edit can assign user
User: passwordConfirm added to entity
User: add/edit only shows userless websites
User: boolean active (true/false)
Tickets: edit show history

+++ src/main/java/com/example/application/data/service/TeamService.java
@@ -2,6 +2,8 @@ package com.example.application.data.service;
 
 import com.example.application.data.entity.*;
 import com.example.application.data.repository.*;
+import com.vaadin.flow.component.notification.Notification;
+import com.vaadin.flow.component.notification.NotificationVariant;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -25,8 +27,22 @@ public class TeamService {
         }
     }
 
+    public Team findTeamByName(String stringFilter){
+            return teamRepository.getTeamByName(stringFilter);
+    }
+
     public void deleteTeam(Team team) {
-        teamRepository.delete(team);
+        if (!team.getWebsites().isEmpty()) {
+            Notification notification = Notification
+                    .show("Team cannot be deleted! Please control if there are websites assigned to this team.");
+            notification.addThemeVariants(NotificationVariant.LUMO_ERROR);
+        }
+        else {
+            teamRepository.delete(team);
+            Notification notification = Notification
+                    .show("Team deleted!");
+            notification.addThemeVariants(NotificationVariant.LUMO_SUCCESS);
+        }
     }
 
     public void saveTeam(Team team) {
